Reorder lifecycle callbacks, fix tests and move some templating tests to a better spot. Remove unnecessary tests.
