+++ test/unit.js
@@ -15,10 +15,9 @@ import './unit/lifecycle-scenarios';
 import './unit/lifecycle/created';
 import './unit/lifecycle/events';
 import './unit/lifecycle/properties';
-import './unit/lifecycle/template';
+import './unit/lifecycle/ready';
 import './unit/registration';
 import './unit/registry';
-import './unit/templating';
 import './unit/util/create-from-html';
 import './unit/util/dash-case';
 import './unit/util/walk-tree';
